Add:  현재 로그인 유저가 리뷰 작성자인지 확인

+++ src/main/java/com/openbook/openbook/event/service/common/CommonEventReviewService.java
@@ -67,6 +67,10 @@ public class CommonEventReviewService {
 
     @Transactional
     public void deleteReview(long userId, long reviewId) {
+        EventReview review = eventReviewService.getEventReviewOrException(reviewId);
+        if(review.getReviewer().getId()!=userId) {
+            throw new OpenBookException(ErrorCode.FORBIDDEN_ACCESS);
+        }
         eventReviewService.deleteEventReview(reviewId);
     }
